Refactor importance evaluator tests: merge tree_based_evaluator and consolidate conditional tests
